# cmd/docs/swagger.yaml
definitions:
  exceptions.Error:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
  handlers.AddSongPayload:
    properties:
      group:
        type: string
      song:
        type: string
    type: object
  models.Song:
    description: Структура, содержащая данные о песне, такие как группа, название
      песни, текст, дата выпуска и ссылка.
    properties:
      group:
        type: string
      id:
        type: integer
      link:
        type: string
      release_date:
        type: string
      song:
        type: string
      text:
        type: string
    type: object
  services.SongByIdWithVersePagination:
    properties:
      current_verse:
        type: integer
      song:
        $ref: '#/definitions/models.Song'
      verses_amount:
        type: integer
    type: object
  services.SongsWithPagination:
    properties:
      current_page:
        type: integer
      pages_amount:
        type: integer
      result:
        items:
          $ref: '#/definitions/models.Song'
        type: array
    type: object
info:
  contact: {}
paths:
  /songs:
    get:
      consumes:
      - application/json
      description: Возвращает информацию о песне по указанному ID с поддержкой пагинации
        (если она нужна).
      parameters:
      - description: Страница
        in: query
        name: page
        type: integer
      - description: Количество элементов
        in: query
        name: limit
        type: integer
      - description: Полнотекстовый поиск по всем полям сущности Song
        in: query
        name: search_query
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Успешный ответ, песня найдена
          schema:
            items:
              $ref: '#/definitions/services.SongsWithPagination'
            type: array
        "400":
          description: Неверный запрос, ID песни не предоставлен или некорректен
          schema:
            $ref: '#/definitions/exceptions.Error'
        "404":
          description: Песня с предоставленным ID не найдена
          schema:
            $ref: '#/definitions/exceptions.Error'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/exceptions.Error'
      summary: Получение всех песен с пагинацией
      tags:
      - songs
    patch:
      consumes:
      - application/json
      description: Обновляет запись о песне
      parameters:
      - description: Данные песни для обновления
        in: body
        name: song
        required: true
        schema:
          $ref: '#/definitions/models.Song'
      produces:
      - application/json
      responses:
        "201":
          description: Песня успешно обновлена
          schema:
            $ref: '#/definitions/models.Song'
        "400":
          description: Некорректный запрос, неправильный формат данных
          schema:
            $ref: '#/definitions/exceptions.Error'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/exceptions.Error'
      summary: Обновление данных песни
      tags:
      - songs
    post:
      consumes:
      - application/json
      description: Создает новую запись о песне
      parameters:
      - description: Данные песни для добавления
        in: body
        name: song
        required: true
        schema:
          $ref: '#/definitions/handlers.AddSongPayload'
      produces:
      - application/json
      responses:
        "201":
          description: Песня успешно добавлена
          schema:
            type: object
        "400":
          description: Некорректный запрос, неправильный формат данных
          schema:
            $ref: '#/definitions/exceptions.Error'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/exceptions.Error'
      summary: Добавление новой песни
      tags:
      - songs
  /songs/{id}:
    delete:
      consumes:
      - application/json
      description: Удаляет песню по ID
      parameters:
      - description: ID песни
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Песня успешно удалена
          schema:
            type: object
        "400":
          description: Некорректный запрос, неправильный формат данных
          schema:
            $ref: '#/definitions/exceptions.Error'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/exceptions.Error'
      summary: Удаление песни
      tags:
      - songs
    get:
      consumes:
      - application/json
      description: Возвращает информацию о песне по указанному ID
      parameters:
      - description: ID песни
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Песня успешно найдена
          schema:
            $ref: '#/definitions/services.SongByIdWithVersePagination'
        "400":
          description: Неверный запрос, ID песни не предоставлен или некорректен
          schema:
            $ref: '#/definitions/exceptions.Error'
        "404":
          description: Песня с предоставленным ID не найдена
          schema:
            $ref: '#/definitions/exceptions.Error'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/exceptions.Error'
      summary: Получение песни по ID
      tags:
      - songs
  /songs/paginated/{id}:
    get:
      consumes:
      - application/json
      description: Возвращает информацию о песне по указанному ID с поддержкой пагинации
        (если она нужна).
      parameters:
      - description: ID песни
        in: path
        name: id
        required: true
        type: integer
      - description: Страница(порядковый номер куплета)
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Песня успешно найдена
          schema:
            $ref: '#/definitions/services.SongByIdWithVersePagination'
        "400":
          description: Неверный запрос, ID песни не предоставлен или некорректен
          schema:
            $ref: '#/definitions/exceptions.Error'
        "404":
          description: Песня с предоставленным ID не найдена
          schema:
            $ref: '#/definitions/exceptions.Error'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/exceptions.Error'
      summary: Получение песни по ID с пагинацией по куплетам
      tags:
      - songs
swagger: "2.0"